requirements.txt
python-telegram-bot
httpx
httpcore
anyio
h11
idna
sniffio
certifi
web3
solana
solders
python-dotenv
websockets
construct-typing
typing-extensions
